requirements.txt
streamlit==1.32.0
pandas==2.2.1
google-generativeai==0.4.1
xlsxwriter==3.1.9